+++ tests/geophires_x_tests/test_options_list.py
@@ -1,3 +1,5 @@
+from typing import ClassVar
+
 from base_test_case import BaseTestCase
 from geophires_x.OptionList import EndUseOptions
 from geophires_x.OptionList import PlantType
@@ -33,9 +35,33 @@ class EndUseOptionsTestCase(BaseTestCase):
 
 
 class WellDrillingCostCorrelationTestCase(BaseTestCase):
+
     def test_equality(self):
         self.assertFalse(WellDrillingCostCorrelation.VERTICAL_SMALL == WellDrillingCostCorrelation.DEVIATED_SMALL)
         self.assertTrue(WellDrillingCostCorrelation.VERTICAL_SMALL == WellDrillingCostCorrelation.VERTICAL_SMALL)
+
+    COST_CORRELATION_TEST_CASES: ClassVar[list] = [
+        (WellDrillingCostCorrelation.VERTICAL_SMALL, 3500, 5.1),
+        (WellDrillingCostCorrelation.VERTICAL_SMALL, 6500, 13.9),
+        (WellDrillingCostCorrelation.VERTICAL_SMALL, 7000, 15.9),
+        (WellDrillingCostCorrelation.VERTICAL_LARGE, 6500, 17.2),
+        (WellDrillingCostCorrelation.DEVIATED_SMALL, 6500, 14.9),
+        (WellDrillingCostCorrelation.DEVIATED_LARGE, 6500, 18.3),
+    ]
+
+    def test_drilling_cost_curve_correlations(self):
+        """
+        Check calibration with NREL 2025 Cost Curve Update.
+        Values in COST_CORRELATION_TEST_CASES derived from graphs on p. 8 of
+        https://pangea.stanford.edu/ERE/db/GeoConf/papers/SGW/2025/Akindipe.pdf?t=1740084555
+        """
+
+        for test_case in WellDrillingCostCorrelationTestCase.COST_CORRELATION_TEST_CASES:
+            correlation: WellDrillingCostCorrelation = test_case[0]
+            depth_m = test_case[1]
+            expected_cost_musd = test_case[2]
+            with self.subTest(msg=str(f'{correlation.name}, {depth_m}m')):
+                self.assertAlmostEqual(expected_cost_musd, correlation.calculate_cost_MUSD(depth_m), delta=0.1)
 
 
 class PlantTypeTestCase(BaseTestCase):